[P6.6] feat(ui): the episode-image panels do what they look like they do
Three surfaces came off the «Kommer» gate at once, because they had one cause
between them and [P6.5] removed it.

THE SHIM. Six methods, previously `async () => ({ ok: false })`. The picker
lives here rather than in the panel, because `thumbnail-panel.ts` calls
`setDefault()` / `setEpisode(rp)` with no path and reads `null` as "user
cancelled — leave the image alone". The picker's filter is the shared image
list MINUS gif: an animated overlay logo is reasonable, an animated podcast
cover is not something any RSS consumer accepts, and the backend refuses the
bytes anyway — this only spares the round trip.

Failures are reshaped into the `{ error: code }` member of the union the panel
actually reads. The backend says `AppError::Validation("too_large")`, which
arrives as `"validation: too_large"`; the panel's `errorLabel()` matches the
BARE code and echoes anything else verbatim, so passing the prefixed string
through would have printed «validation: too_large» at the user instead of
«Filen er for stor (over 20 MB)».

THE TYPES. `ThumbnailInfo`/`ThumbnailResult`/`ThumbnailResolved` were
hand-written in main.ts for a backend that did not exist. They are now the
generated ts-rs bindings, so a change to the Rust DTO is a tsc error rather
than a panel reading a field that stopped being sent.

THE HOME CARD. It read `settings.defaultThumbnailPath` — a field NOTHING has
ever written in the Tauri build. So it appeared only for users carrying a stale
path from Electron, showed a file nothing consumed, said «Episodebilde kommer —
brukes ikke ennå», and disabled its own «Endre» because that led to a gated
panel. It now asks the backend what the default is, shows it with its pixel
size, and lets you change it. The phantom settings field is gone too — it
described a `<name>.thumb.{ext}` layout that contradicts the real one.

`applyComingSoonGate` is deleted. It had exactly three callers, all three were
these panels, and its own doc-comment ("no Rust side exists at all") had become
false. Keeping a ready-made «Kommer» chip around is an invitation to reach for
the chip instead of writing the missing half. Its two locale keys went with it,
plus the dead `home.thumbComing`.

The stretch («Hent fra video») is deliberately not in this commit — see the
report; the six-command family and its three surfaces are finished and green,
and that is the thing worth shipping.

+++ src-tauri/src/commands/thumbnail.rs
@@ -93,7 +93,11 @@ pub enum ThumbnailKind {
 pub struct ThumbnailInfo {
     pub width: u32,
     pub height: u32,
-    pub byte_size: u64,
+    /// `u32`, not `u64`, on purpose: [`MAX_BYTES`] caps this at 20 MB, and ts-rs
+    /// maps `u64` to `bigint` — which the panel's `byteSize > 5 * 1024 * 1024`
+    /// comparison and its `fmtBytes(b: number)` would both reject, for a value
+    /// that arrives over JSON IPC as an ordinary number anyway.
+    pub byte_size: u32,
     pub format: ImageFormat,
 }
 
@@ -196,7 +200,7 @@ fn view_of(path: &Path, kind: Option<ThumbnailKind>) -> AppResult<ThumbnailView>
         info: ThumbnailInfo {
             width: probe.width,
             height: probe.height,
-            byte_size: bytes.len() as u64,
+            byte_size: bytes.len() as u32,
             format,
         },
         data_url: format!(
@@ -501,7 +505,7 @@ mod tests {
         let v = set_episode(&rec, &write(&media, "art.png", &bytes)).unwrap();
 
         assert!(v.data_url.starts_with("data:image/png;base64,"));
-        assert_eq!(v.info.byte_size, bytes.len() as u64);
+        assert_eq!(v.info.byte_size, bytes.len() as u32);
         // …and it really is the file, not a placeholder.
         use base64::{engine::general_purpose::STANDARD, Engine as _};
         let encoded = v.data_url.split_once(",").unwrap().1;
